Make it compile in modern C.
Use some modern types, and use more specific types than `char*` (mostly, the
authors meant `struct fb*`, so use that).

Note a significant change:

-logical        slop    2;
+size_t         slop = sizeof(void*);

2 was the word size on the machine for which the code was written, and the code
assumes that is the right minimum value for `slop`. This change makes `slop`'s
value the word size on the target machine, whatever that may be.